orchestrator/requirements.txt
httpx>=0.27.0
pydantic>=2.6.1
pyyaml>=6.0.1
python-dotenv>=1.0.1
structlog>=24.1.0
colorama>=0.4.6
pyautogui>=0.9.54
pywinauto>=0.6.9
pyperclip>=1.8.2
Pillow>=10.0.0
